add boss room template
it is a long room with thick walls. Player spawn area is placed at the left; exit area (could be mapped to a credits scene/fixed text?) is on the right. There is plenty of space in the middle to place a boss.

As of now (12:27 am) this room has not been integrated with the RoomManager yet.